+++ commands/replay.js
@@ -5,16 +5,18 @@ module.exports = {
   name: "replay",
   desc: "Reset the current song progress.",
   args: [],
-  exec: ({Emojis, Queue, Voice}, Message, Args) => {
-    if (Voice.Handler._destroyed) {
-      Voice.Handler = null;
+  exec: ({Emojis, Server}, Message, Args) => {
+    const { Queue, Voice } = Server;
+    
+    if (Voice._destroyed) {
+      Server.Voice = null;
     }
     
-    if (!Voice.Handler) return Message.channel.send(`${Emojis.WARNING} No music is playing !`);
+    if (!Voice) return Message.channel.send(`${Emojis.WARNING} No music is playing !`);
 
     if (Queue.isEmpty()) return Message.channel.send(`${Emojis.WARNING} The queue is empty !`);
-
-    Voice.Handler.restart();
+    
+    Voice.restart();
     Message.channel.send(`${Emojis.MUSIC_RESTART} **Restarted** song`);
   }
 }