+++ events/guildCreate.js
@@ -1,3 +1,6 @@
+/** Queue system. */
+const Queue = require("../lib/Queue.js").Queue;
+
 const updateActivity = require("../lib/Utils.js").updateActivity;
 
 /**
@@ -10,4 +13,8 @@ module.exports = (Client, [Guild]) => {
   updateActivity(Client).then(presence => {
     console.log(`INFO: Activity updated.`);
   }).catch(console.error);
+
+  /** Initiate queue and voice. */
+  Client.guilds.get(Guild.id).Queue = new Queue();
+  Client.guilds.get(Guild.id).Voice = null;
 };